+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Cyber-AutoAgent-ng Changelog
 
+- Configure Ollama keep-alive for models to avoid extra start up time. Defaults to 30m.
 - Add bug bounty header markers. (Fixes #63)
 - Add idor_specialist tool. (Fixes #22)
 - Publish tools image to `public.ecr.aws/bramblethorn/cyber-autoagent-ng/tools:latest`. (Fixes #20)